Arrange folder/file structure so that gcc/nvcc cross compile is supported
